:necktie: modify like, want business logic

+++ back/src/api/routes/want.js
@@ -18,9 +18,10 @@ router.get('/', async (req, res) => {
 
 router.post('/', async (req, res) => {
     const artwork_id = req.artwork_id;
-    const { user_id } = req.query;
+    const { user_id } = req.session;
 
     try{
+        if(!user_id) return res.status(401).json('Nonauthorized');
         const want = await WantServiceInstance.postArtworkWant(artwork_id, user_id);
         res.status(201).json(want);
     }
@@ -31,9 +32,10 @@ router.post('/', async (req, res) => {
 
 router.delete('/', async (req, res) => {
     const artwork_id = req.artwork_id;
-    const { user_id } = req.query;
-
+    const { user_id } = req.session;
+    
     try{
+        if(!user_id) return res.status(401).json('Nonauthorized');
         const want = await WantServiceInstance.deleteArtworkWant(artwork_id, user_id);
         res.status(201).json(want);
     }
